Move prompt types to common project.
