+++ DaprAI.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DaprAI.Application", "src\D
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DaprAI.Components", "src\DaprAI.Components\DaprAI.Components.csproj", "{989CDF22-9CE6-49FA-BDEC-043E8D83094C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DaprAI.Common", "src\DaprAI.Common\DaprAI.Common.csproj", "{7C8070D8-2D73-4852-8791-D0167AC914BD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,9 +28,14 @@ Global
 		{989CDF22-9CE6-49FA-BDEC-043E8D83094C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{989CDF22-9CE6-49FA-BDEC-043E8D83094C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{989CDF22-9CE6-49FA-BDEC-043E8D83094C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C8070D8-2D73-4852-8791-D0167AC914BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C8070D8-2D73-4852-8791-D0167AC914BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C8070D8-2D73-4852-8791-D0167AC914BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C8070D8-2D73-4852-8791-D0167AC914BD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{9650A96A-527C-4BB1-82F3-3911987D6266} = {CC75084B-F654-4A77-BE33-D35C25C79FA8}
 		{989CDF22-9CE6-49FA-BDEC-043E8D83094C} = {CC75084B-F654-4A77-BE33-D35C25C79FA8}
+		{7C8070D8-2D73-4852-8791-D0167AC914BD} = {CC75084B-F654-4A77-BE33-D35C25C79FA8}
 	EndGlobalSection
 EndGlobal
