+++ src/DaprAI.Application/Program.cs
@@ -1,4 +1,4 @@
-using System.Text.Json.Serialization;
+using DaprAI;
 using Dapr.Client;
 using Microsoft.AspNetCore.Mvc;
 
@@ -49,39 +49,4 @@ app.MapPost(
     .WithName("Summarize")
     .WithOpenApi();
 
-app.MapPost($"/{AzureAI}", () =>
-{
-    // TODO: Handle chat response.
-    return Results.Accepted();
-});
-
-app.MapPost($"/{AzureOpenAI}", () =>
-{
-    // TODO: Handle chat response.
-    return Results.Accepted();
-});
-
-app.MapPost($"/{ChatGpt}", () =>
-{
-    // TODO: Handle chat response.
-    return Results.Accepted();
-});
-
 app.Run();
-
-internal sealed record PromptRequest(
-    [property: JsonPropertyName("prompt")]
-    string Prompt);
-
-internal sealed record PromptResponse(
-    [property: JsonPropertyName("response")]
-    string Response);
-
-internal static class DaprClientExtensions
-{
-    public static Task<PromptResponse> PromptAIAsync(this DaprClient daprClient, string component, PromptRequest request) =>
-        daprClient.InvokeBindingAsync<PromptRequest, PromptResponse>(component, "prompt", request);
-
-    public static Task<PromptResponse> SummarizeAIAsync(this DaprClient daprClient, string component, PromptRequest request) =>
-        daprClient.InvokeBindingAsync<PromptRequest, PromptResponse>(component, "summarize", request);
-}
