+++ src/DaprAI.Application/components/azure-open-ai-binding.yaml
@@ -1,0 +1,22 @@
+apiVersion: dapr.io/v1alpha1
+kind: Component
+metadata:
+  name: azure-open-ai
+spec:
+  type: bindings.azure-open-ai
+  version: v1
+  metadata:
+    - name: deployment
+      secretKeyRef:
+        name: azure-open-ai-deployment
+        key: azure-open-ai-deployment
+    - name: endpoint
+      secretKeyRef:
+        name: azure-open-ai-endpoint
+        key: azure-open-ai-endpoint
+    - name: key
+      secretKeyRef:
+        name: azure-open-ai-key
+        key: azure-open-ai-key
+auth:
+  secretStore: secrets
