+++ src/DaprAI.Application/components/chat-gpt-binding.yaml
@@ -6,9 +6,11 @@ spec:
   type: bindings.chat-gpt
   version: v1
   metadata:
-    - name: open-api-endpoint
-      value: https://api.openai.com
-    - name: open-api-key
+    - name: endpoint
+      secretKeyRef:
+        name: open-api-endpoint
+        key: open-api-endpoint
+    - name: key
       secretKeyRef:
         name: open-api-key
         key: open-api-key
