+++ src/DaprAI.Common/PromptResponse.cs
@@ -2,9 +2,9 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace DaprAI.Bindings;
+namespace DaprAI;
 
-internal sealed record PromptResponse(
+public sealed record PromptResponse(
     [property: JsonPropertyName("response")]
     string Response)
 {
