+++ src/DaprAI.Components/DaprAI.Components.csproj
@@ -7,6 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\DaprAI.Common\DaprAI.Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Azure.AI.TextAnalytics" Version="5.3.0-beta.1" />
     <PackageReference Include="Dapr.PluggableComponents.AspNetCore" Version="0.1.0" />
   </ItemGroup>
